backend/requirements.txt
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
asttokens==3.0.0
attrs==25.4.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
cloudinary==1.44.1
colorama==0.4.6
crudcleanarch @ git+https://github.com/ZequerVirus/CleanCrud.git@e15a3e887b715c0f414d0be052e3ad26c4a34c2b
dateparser==1.2.2
decorator==5.2.1
distro==1.9.0
Django==5.2.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-decouple==2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
executing==2.2.1
filelock==3.20.0
frozenlist==1.8.0
future==1.0.0
gevent==25.9.1
google-auth==2.43.0
google-genai==1.49.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipython==8.37.0
jedi==0.19.2
jiter==0.11.1
joblib==1.5.2
matplotlib-inline==0.2.1
multidict==6.7.0
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2==2.9.11
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2025.10.23
reportlab==4.4.4
requests==2.32.5
requests-file==3.0.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
stripe==13.0.1
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tldextract==5.3.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
wcwidth==0.2.14
websockets==15.0.1
wheel==0.45.1
yarl==1.22.0
zope.event==6.0
zope.interface==8.0.1
